requirements.txt
langchain-google-genai
langgraph
langchain_community
wikipedia
python-dotenv
duckduckgo-search